requirements.txt
python-dotenv
langchain
langchain-community
langchain-openai
openai
tiktoken
chromadb